fix: elimino prints, la banana pierde municion, la granada gira

+++ server/logic/duck.cpp
@@ -371,7 +371,7 @@ PlayerDTO Duck::toDTO() const {
 
 Duck::~Duck() {}
 
-uint8_t Duck::getWins() const { return wins; }
+int Duck::getWins() const { return wins; }
 
 void Duck::setWins() { wins = 200; }
